Add SourceData interface and modified DatalakeFile to use it
